+++ src/benders.rs
@@ -13,38 +13,108 @@ use std::collections::HashMap;
 /// 
 /// Represents the entire TOML options file.
 #[derive(Debug, Deserialize, MutConfig)]
-#[allow(unused_attributes)] // pops up a warning for custom attributes apparently.
+#[allow(unused_attributes, non_snake_case)] // pops up a warning for custom attributes apparently.
 pub struct MainConfig {
+    /// The name of the input file.
+    /// Is required.
     #[ignore]
-    inputfile : String, // Input file.
+    inputfile : String,
+
+    /// The name of the output file.
+    /// If not passed, the name of the input file is used.
     #[ignore]
-    outputfile : Option<String>, // Manually setting the output file.
+    outputfile : Option<String>,
+
+    /// The number of times to repeat the program.
+    /// If not passed, the value defaults to 1.
     #[ignore]
     pub times : Option<isize>,
-    iterations: Vec<isize>, // How many iteration every "mutate" does
-    chunksize: Vec<isize>, // A range of chunksizes.
+
+    /// A range of iterations.
+    /// Specifies how many times each mutation is applied.
+    /// A global option to be set for all relevant mutations.
+    iterations: Vec<isize>,
+
+    /// A range of chunksizes.
+    /// Specifies the size of each chunk of *bytes* to mutate.
+    /// A global option to be set for all relevant mutations.
+    chunksize: Vec<isize>,
+
+    /// A list of mutations to be used. 
     #[ignore]
     pub mutations: Vec<Vec<String>>,
-    loop_mut: LoopConfig,
-    increase_mut: IncConfig,
-    gradient_mut: GraConfig
+
+    // Mutation configurations.
+    VoidConfig: Option<VoidConfig>,
+    ChaosConfig: Option<ChaosConfig>,
+    LoopConfig: Option<LoopConfig>,
+    ReverseConfig: Option<ReverseConfig>,
+    ShiftConfig: Option<ShiftConfig>,
+    ShuffleConfig: Option<ShuffleConfig>,
+    SwapConfig: Option<SwapConfig>,
+    IncreaseConfig: Option<IncreaseConfig>,
+    GradientConfig: Option<GradientConfig>
 }
 
-#[derive(Debug, Deserialize, MutConfig)]
+#[derive(Debug, Deserialize, Clone, MutConfig)]
+pub struct VoidConfig {
+    iterations: Option<Vec<isize>>,
+    chunksize: Option<Vec<isize>>,
+}
+
+#[derive(Debug, Deserialize, Clone, MutConfig)]
+pub struct ChaosConfig {
+    iterations: Option<Vec<isize>>,
+    chunksize: Option<Vec<isize>>,
+}
+
+#[derive(Debug, Deserialize, Clone, MutConfig)]
 pub struct LoopConfig {
-    loops: Vec<isize>
+    iterations: Option<Vec<isize>>,
+    chunksize: Option<Vec<isize>>,
+    loops: Option<Vec<isize>>
 }
 
-#[derive(Debug, Deserialize, MutConfig)]
-pub struct IncConfig {
-    increase_by: Vec<isize>
+#[derive(Debug, Deserialize, Clone, MutConfig)]
+pub struct ReverseConfig {
+    iterations: Option<Vec<isize>>,
+    chunksize: Option<Vec<isize>>,
 }
 
-#[derive(Debug, Deserialize, MutConfig)]
-pub struct GraConfig {
-    accelerate_by: Vec<isize>,
-    accelerate_in: Vec<isize> 
+#[derive(Debug, Deserialize, Clone, MutConfig)]
+pub struct ShiftConfig {
+    iterations: Option<Vec<isize>>,
+    chunksize: Option<Vec<isize>>,
 }
+
+#[derive(Debug, Deserialize, Clone, MutConfig)]
+pub struct ShuffleConfig {
+    iterations: Option<Vec<isize>>,
+    chunksize: Option<Vec<isize>>,
+}
+
+#[derive(Debug, Deserialize, Clone, MutConfig)]
+pub struct SwapConfig {
+    iterations: Option<Vec<isize>>,
+    chunksize: Option<Vec<isize>>,
+}
+
+#[derive(Debug, Deserialize, Clone, MutConfig)]
+pub struct IncreaseConfig {
+    iterations: Option<Vec<isize>>,
+    chunksize: Option<Vec<isize>>,
+    increase_by: Option<Vec<isize>>
+}
+
+#[derive(Debug, Deserialize, Clone, MutConfig)]
+pub struct GradientConfig {
+    iterations: Option<Vec<isize>>,
+    chunksize: Option<Vec<isize>>,
+    accelerate_by: Option<Vec<isize>>,
+    accelerate_in: Option<Vec<isize>> 
+}
+
+
 
 /// A main controller of the databender.
 /// 
@@ -72,8 +142,263 @@ impl KaBender {
             log : Vec::new(),
         };
 
+        new.setup_config();
         new.init_file();
         new
+    }
+
+    /// Sets up the configurations.
+    /// 
+    /// Is a huge function due to multiple repeated boilerplate code.
+    /// In the future, there might be macros etc. to improve the style.
+    fn setup_config(&mut self) {
+        let muts_passed = self.config.mutations.concat();
+
+        self.config.iterations = verify_num_option(&self.config.iterations, "iterations", "globally");
+        self.config.chunksize = verify_num_option(&self.config.chunksize, "chunksize", "globally");
+
+        let void_exists = muts_passed.contains(&String::from("Void"));
+        let chaos_exists = muts_passed.contains(&String::from("Chaos"));
+        let loops_exists = muts_passed.contains(&String::from("Loops"));
+        let reverse_exists = muts_passed.contains(&String::from("Reverse"));
+        let shift_exists = muts_passed.contains(&String::from("Shift"));
+        let shuffle_exists = muts_passed.contains(&String::from("Shuffle"));
+        let swap_exists = muts_passed.contains(&String::from("Swap"));
+        let increase_exists = muts_passed.contains(&String::from("Increase"));
+        let gradient_exists = muts_passed.contains(&String::from("Gradient"));
+
+        // If mutation not included, reset it to None.
+        if !void_exists         {self.config.VoidConfig         = None};
+        if !chaos_exists        {self.config.ChaosConfig        = None};
+        if !loops_exists        {self.config.LoopConfig         = None};
+        if !reverse_exists      {self.config.ReverseConfig      = None};
+        if !shift_exists        {self.config.ShiftConfig        = None};
+        if !shuffle_exists      {self.config.ShuffleConfig      = None};
+        if !swap_exists         {self.config.SwapConfig         = None};
+        if !increase_exists     {self.config.IncreaseConfig     = None};
+        if !gradient_exists     {self.config.GradientConfig     = None};
+
+        // VoidConfig setup
+        if self.config.VoidConfig.is_none() && void_exists {
+            self.config.VoidConfig = Some(VoidConfig {
+                iterations: Some(self.config.iterations.clone()),
+                chunksize: Some(self.config.chunksize.clone())
+            });
+        }
+        else if let (Some(x), true) = (&mut self.config.VoidConfig, void_exists) {
+            x.chunksize = if let Some(ch) = &x.chunksize
+                {Some(verify_num_option(&ch, "chunksize", "under '[VoidConfig]'"))}
+                else {Some(self.config.chunksize.clone())};
+            
+            x.iterations = if let Some(ch) = &x.iterations
+                {Some(verify_num_option(&ch, "iterations", "under '[VoidConfig]'"))}
+                else {Some(self.config.iterations.clone())};
+        };
+
+        // ChaosConfig setup
+        if self.config.ChaosConfig.is_none() && chaos_exists {
+            self.config.ChaosConfig = Some(ChaosConfig {
+                iterations: Some(self.config.iterations.clone()),
+                chunksize: Some(self.config.chunksize.clone())
+            });
+        }
+        else if let (Some(x), true) = (&mut self.config.ChaosConfig, chaos_exists) {
+            x.chunksize = if let Some(ch) = &x.chunksize
+                {Some(verify_num_option(&ch, "chunksize", "under '[ChaosConfig]'"))}
+                else {Some(self.config.chunksize.clone())};
+            
+            x.iterations = if let Some(ch) = &x.iterations
+                {Some(verify_num_option(&ch, "iterations", "under '[ChaosConfig]'"))}
+                else {Some(self.config.iterations.clone())};
+        };
+
+        // LoopConfig setup
+        if self.config.LoopConfig.is_none() && loops_exists {
+            let example = r#"
+                [LoopConfig]
+                loops = [1,2]
+            "#;
+
+            panic!("You have added a 'Loops' mutation, but haven't passed its options.
+                    \nSpecifically, the 'loops' option needs to be passed under '[LoopConfig]'.
+                    \nThe following is an example:
+                    \n{}", example);
+        }
+        else if let (Some(x), true) = (&mut self.config.LoopConfig, loops_exists) {
+            x.chunksize = if let Some(ch) = &x.chunksize
+                {Some(verify_num_option(&ch, "chunksize", "under '[LoopConfig]'"))}
+                else {Some(self.config.chunksize.clone())};
+            
+            x.iterations = if let Some(ch) = &x.iterations
+                {Some(verify_num_option(&ch, "iterations", "under '[LoopConfig]'"))}
+                else {Some(self.config.iterations.clone())};
+
+            if x.loops.is_none() {
+                panic!("You have added a 'Loops' mutation, but haven't passed the 'loops' option.
+                        \nFor example: 'loops = [1,5]'");
+            }
+            else if let Some(l) = &x.loops {
+                x.loops = Some(verify_num_option(&l, "loops", "under '[LoopConfig]'"));
+            };
+        };
+
+        // ReverseConfig setup
+        if self.config.ReverseConfig.is_none() && reverse_exists {
+            self.config.ReverseConfig = Some(ReverseConfig {
+                iterations: Some(self.config.iterations.clone()),
+                chunksize: Some(self.config.chunksize.clone())
+            });
+        }
+        else if let (Some(x), true) = (&mut self.config.ReverseConfig, reverse_exists) {
+            x.chunksize = if let Some(ch) = &x.chunksize
+                {Some(verify_num_option(&ch, "chunksize", "under '[ReverseConfig]'"))}
+                else {Some(self.config.chunksize.clone())};
+            
+            x.iterations = if let Some(ch) = &x.iterations
+                {Some(verify_num_option(&ch, "iterations", "under '[ReverseConfig]'"))}
+                else {Some(self.config.iterations.clone())};
+        };
+
+        // ShiftConfig setup
+        if self.config.ShiftConfig.is_none() && shift_exists {
+            self.config.ShiftConfig = Some(ShiftConfig {
+                iterations: Some(self.config.iterations.clone()),
+                chunksize: Some(self.config.chunksize.clone())
+            });
+        }
+        else if let (Some(x), true) = (&mut self.config.ShiftConfig, shift_exists) {
+            x.chunksize = if let Some(ch) = &x.chunksize
+                {Some(verify_num_option(&ch, "chunksize", "under '[ShiftConfig]'"))}
+                else {Some(self.config.chunksize.clone())};
+            
+            x.iterations = if let Some(ch) = &x.iterations
+                {Some(verify_num_option(&ch, "iterations", "under '[ShiftConfig]'"))}
+                else {Some(self.config.iterations.clone())};
+        };
+
+        // ShuffleConfig setup
+        if self.config.ShuffleConfig.is_none() && shuffle_exists {
+            self.config.ShuffleConfig = Some(ShuffleConfig {
+                iterations: Some(self.config.iterations.clone()),
+                chunksize: Some(self.config.chunksize.clone())
+            });
+        }
+        else if let (Some(x), true) = (&mut self.config.ShuffleConfig, shuffle_exists) {
+            x.chunksize = if let Some(ch) = &x.chunksize
+                {Some(verify_num_option(&ch, "chunksize", "under '[ShuffleConfig]'"))}
+                else {Some(self.config.chunksize.clone())};
+            
+            x.iterations = if let Some(ch) = &x.iterations
+                {Some(verify_num_option(&ch, "iterations", "under '[ShuffleConfig]'"))}
+                else {Some(self.config.iterations.clone())};
+        };
+
+        // SwapConfig setup
+        if self.config.SwapConfig.is_none() && swap_exists {
+            self.config.SwapConfig = Some(SwapConfig {
+                iterations: Some(self.config.iterations.clone()),
+                chunksize: Some(self.config.chunksize.clone())
+            });
+        }
+        else if let (Some(x), true) = (&mut self.config.SwapConfig, swap_exists) {
+            x.chunksize = if let Some(ch) = &x.chunksize
+                {Some(verify_num_option(&ch, "chunksize", "under '[SwapConfig]'"))}
+                else {Some(self.config.chunksize.clone())};
+            
+            x.iterations = if let Some(ch) = &x.iterations
+                {Some(verify_num_option(&ch, "iterations", "under '[SwapConfig]'"))}
+                else {Some(self.config.iterations.clone())};
+        };
+
+        // IncreaseConfig setup
+        if self.config.IncreaseConfig.is_none() && increase_exists {
+            let example = r#"
+                [IncreaseConfig]
+                increase_by = [1,2]
+            "#;
+
+            panic!("You have added a 'IncreaseConfig' mutation, but haven't passed its options.
+                    \nSpecifically, the 'increase_by' option needs to be passed under '[IncreaseConfig]'.
+                    \nThe following is an example:
+                    \n{}", example);
+        }
+        else if let (Some(x), true) = (&mut self.config.IncreaseConfig, increase_exists) {
+            x.chunksize = if let Some(ch) = &x.chunksize
+                {Some(verify_num_option(&ch, "chunksize", "under '[IncreaseConfig]'"))}
+                else {Some(self.config.chunksize.clone())};
+            
+            x.iterations = if let Some(ch) = &x.iterations
+                {Some(verify_num_option(&ch, "iterations", "under '[IncreaseConfig]'"))}
+                else {Some(self.config.iterations.clone())};
+
+            if x.increase_by.is_none() {
+                panic!("You have added a 'Increase' mutation, but haven't passed the 'increase_by' option.
+                        \nFor example: 'increase_by = [1,5]'");
+            }
+            else if let Some(l) = &x.increase_by {
+                x.increase_by = Some(verify_num_option(&l, "increase_by", "under '[LoopConfig]'"));
+            };
+        };
+
+        // GradientConfig setup
+        if self.config.GradientConfig.is_none() && gradient_exists {
+            let example = r#"
+                [GradientConfig]
+                accelerate_by = [1,2]
+                accelerate_in = [1,2]
+            "#;
+
+            panic!("You have added a 'GradientConfig' mutation, but haven't passed its options.
+                    \nSpecifically, the 'accelerate_by' and 'accelerate_in'
+                    \noption needs to be passed under '[LoopConfig]'.
+                    \nThe following is an example:
+                    \n{}", example);
+        }
+        else if let (Some(x), true) = (&mut self.config.GradientConfig, gradient_exists) {
+            x.chunksize = if let Some(ch) = &x.chunksize
+                {Some(verify_num_option(&ch, "chunksize", "under '[GradientConfig]'"))}
+                else {Some(self.config.chunksize.clone())};
+            
+            x.iterations = if let Some(ch) = &x.iterations
+                {Some(verify_num_option(&ch, "iterations", "under '[GradientConfig]'"))}
+                else {Some(self.config.iterations.clone())};
+
+            if x.accelerate_by.is_none() {
+                panic!("You have added a 'Gradient' mutation, but haven't passed the 'accelerate_by' option.
+                        \nFor example: 'accelerate_by = [1,5]'");
+            }
+            else if let Some(l) = &x.accelerate_by {
+                x.accelerate_by = Some(verify_num_option(&l, "accelerate_by", "under '[GradientConfig]'"));
+            };
+
+            if x.accelerate_in.is_none() {
+                panic!("You have added a 'Gradient' mutation, but haven't passed the 'accelerate_in' option.
+                        \nFor example: 'accelerate_in = [1,5]'");
+            }
+            else if let Some(l) = &x.accelerate_in {
+                x.accelerate_in = Some(verify_num_option(&l, "accelerate_in", "under '[GradientConfig]'"));
+            };
+        };
+
+        fn verify_num_option(v : &Vec<isize>, name: &str, location: &str) -> Vec<isize> {
+            let len = v.len();
+            if len != 1 && len != 2 {
+                panic!("You have passed an invalid '{0}' option {2}. It needs to be a vector with 1
+                        \nor 2 elements. [exact = '{0} = [3]', range = '{0} = [1,5]'
+                        \nYour option: {1:?}", name, v, location);
+            }
+            else if len == 1 {
+                return vec![v[0], v[0]+1];
+            }
+            else {
+                if v[0] > v[1] {
+                    return vec![v[1], v[0]+1];
+                }
+                else {
+                    return vec![v[0], v[1]+1];
+                }
+            }
+        }
     }
 
     /// Initialises the file.